Add a GetMetadataBlockInfo method to HeifImageHandle
This method allows callers to get information about the contents of a
metadata block when enumerating through all of the metadata in the
image.
The HeifMetadataBlockInfo class contains the id, item type,
content type and size in bytes of the specified metadata block.

Also added a GetMetadata overload that takes HeifMetadataBlockInfo.