Simplify UserDescriptionProperty.NativeData
There is no need to call DangerousAddRef and DangerousRelease on the
SafeHandles because the class will already ensure that they remain
valid for its lifetime.

It is up to the caller to ensure that the structure returned by the
CreateNativeStructure method is not used after the class has been
disposed.